convert carrot-traffic-driver to TypeScript
- format file
- import config
- simplify repeated config
